Käyttäjä kirjataan ulos samalla kun profiili poistetaan. Profiilisivu ei näy, jos ei ole kirjautuneena
